fix: extract shared agent_builder helper to eliminate duplication between local.rs and monitor.rs
