+++ crates/jyc-cli/src/cli/mod.rs
@@ -1,3 +1,4 @@
+pub mod agent_builder;
 pub mod config;
 pub mod dashboard;
 pub mod local;
