+++ crates/jyc-cli/src/cli/monitor.rs
@@ -7,8 +7,6 @@ use tokio_util::sync::CancellationToken;
 use tracing::Instrument;
 
 use jyc_agent::JycAgentService;
-use jyc_core::agent::AgentService;
-use jyc_core::static_agent::StaticAgentService;
 
 use jyc_services::job_scheduler::JobScheduler;
 use std::collections::HashMap;
@@ -341,140 +339,20 @@ pub async fn run(args: &MonitorArgs, workdir: &Path) -> Result<()> {
         all_outbounds.push((channel_name.clone(), outbound.clone()));
 
         // Create agent based on configured mode
-        // Resolve effective model per-channel: channel-level override beats global
-        let effective_model = channel_config
-            .model
-            .clone()
-            .or_else(|| agent_config.model.clone());
-        let effective_small_model = channel_config
-            .small_model
-            .clone()
-            .or_else(|| agent_config.small_model.clone());
-        let agent: Arc<dyn AgentService> = match agent_config.mode.as_str() {
-            "agent" => {
-                // In-process agent
-                let model = effective_model;
-                tracing::info!(channel = %channel_name, model = ?model, "Using agent: jyc-agent (in-process)");
-                let providers = agent_config
-                    .providers
-                    .iter()
-                    .map(|(name, def)| {
-                        let models = def
-                            .models
-                            .iter()
-                            .map(|(model_name, model_def)| {
-                                (
-                                    model_name.clone(),
-                                    jyc_agent::types::ModelConfig {
-                                        context_window: model_def.context_window,
-                                        supports_images: model_def.supports_images,
-                                        params: model_def.params.clone(),
-                                        user_agent: model_def.user_agent.clone(),
-                                    },
-                                )
-                            })
-                            .collect();
-                        (
-                            name.clone(),
-                            jyc_agent::types::ProviderConfig {
-                                provider_type: def.provider_type.clone(),
-                                base_url: def.base_url.clone(),
-                                api_key_env: def.api_key_env.clone(),
-                                context_window: def.context_window,
-                                supports_images: def.supports_images,
-                                params: def.params.clone(),
-                                user_agent: def.user_agent.clone(),
-                                models,
-                            },
-                        )
-                    })
-                    .collect();
-                let agent_cfg = jyc_agent::types::AgentConfig {
-                    model,
-                    small_model: effective_small_model.clone(),
-                    providers,
-                    max_iterations: agent_config.max_iterations,
-                    vision: agent_config
-                        .vision
-                        .clone()
-                        .map(|v| jyc_agent::types::VisionConfig {
-                            enabled: v.enabled,
-                            provider: v.provider,
-                            model: v.model,
-                            prompt: v.prompt,
-                        }),
-                };
-                // Use only this channel's patterns so per-pattern overrides
-                // (model, small_model, mcps, disabled_builtin_tools,
-                // inject_inbound_images) are resolved deterministically.
-                // Previously, patterns were flattened from every channel and
-                // find(|p| p.name == name) could return a different channel's
-                // pattern with the same name (HashMap iteration order is
-                // non-deterministic), causing per-pattern overrides to be
-                // silently ignored.
-                let channel_patterns = patterns.clone();
-                // Build optional VisionClient for vision fallback
-                let vision_client: Option<std::sync::Arc<jyc_agent::vision::VisionClient>> = {
-                    agent_config
-                        .vision
-                        .as_ref()
-                        .filter(|v| v.enabled)
-                        .and_then(|v| {
-                            let provider_def = agent_config.providers.get(&v.provider)?;
-                            let base_url = provider_def
-                                .base_url
-                                .clone()
-                                .unwrap_or_else(|| "https://api.deepseek.com".to_string());
-                            let api_key_env = provider_def
-                                .api_key_env
-                                .clone()
-                                .unwrap_or_else(|| "DEEPSEEK_API_KEY".to_string());
-                            let api_key = std::env::var(&api_key_env).unwrap_or_default();
-                            if api_key.is_empty() {
-                                tracing::warn!(
-                                    provider = %v.provider,
-                                    api_key_env = %api_key_env,
-                                    "Vision fallback: API key not found in environment"
-                                );
-                                return None;
-                            }
-                            Some(std::sync::Arc::new(jyc_agent::vision::VisionClient::new(
-                                base_url,
-                                api_key,
-                                v.model.clone(),
-                                v.prompt.clone(),
-                            )))
-                        })
-                };
-                let jyc_agent_svc = Arc::new(JycAgentService::new(
-                    agent_cfg,
-                    workdir.to_path_buf(),
-                    config_snapshot.mcps.clone(),
-                    channel_config.mcps.clone(),
-                    channel_patterns,
-                    inbound_attachment_config.clone(),
-                    vision_client,
-                    Some(outbound.clone()),
-                    channel_config.disabled_tools.clone(),
-                    channel_config.disabled_mcp_servers.clone(),
-                    channel_config.skills.clone(),
-                    channel_config.disabled_skills.clone(),
-                    channel_name.clone(),
-                ));
-                all_agent_services.push(jyc_agent_svc.clone());
-                jyc_agent_svc as Arc<dyn AgentService>
-            }
-            "static" => {
-                let text = agent_config
-                    .text
-                    .as_deref()
-                    .unwrap_or("Thank you for your message.");
-                Arc::new(StaticAgentService::new(text))
-            }
-            other => {
-                anyhow::bail!("unsupported agent mode: '{other}'");
-            }
-        };
+        let agent_result = crate::cli::agent_builder::build_agent_service(
+            &agent_config,
+            channel_config,
+            workdir,
+            outbound.clone(),
+            patterns.clone(),
+            config_snapshot.mcps.clone(),
+            inbound_attachment_config.clone(),
+            channel_name,
+        )?;
+        if let Some(ref jyc_svc) = agent_result.jyc_agent {
+            all_agent_services.push(jyc_svc.clone());
+        }
+        let agent = agent_result.agent;
 
         let template_dir = workdir.join("templates");
